[pymorse] mv poll thread class to stream module
