+++ bindings/pymorse/src/pymorse/pymorse.py
@@ -297,7 +297,7 @@ import asyncore
 import threading
 
 from .future import MorseExecutor
-from .stream import Stream, StreamJSON
+from .stream import Stream, StreamJSON, PollThread
 
 logger = logging.getLogger("pymorse")
 logger.setLevel(logging.WARNING)
@@ -434,20 +434,8 @@ class ResponseCallback:
                 condition.notify_all()
         del ResponseCallback._conditions[:] # clear list
 
-class PollThread(threading.Thread):
-    def __init__(self, timeout=0.01):
-        threading.Thread.__init__(self)
-        self.keep_polling = True
-        self.timeout = timeout
-    def run(self):
-        while asyncore.socket_map and self.keep_polling:
-            asyncore.poll(self.timeout, asyncore.socket_map)
-    def syncstop(self, timeout=None):
-        self.keep_polling = False
-        return self.join(timeout)
-
 class Morse(object):
-    _asyncore_thread = None
+    poll_thread = None
     def __init__(self, host = "localhost", port = 4000):
         """ Creates an instance of the MORSE simulator proxy.
 
@@ -459,9 +447,9 @@ class Morse(object):
         self.host = host
         self.simulator_service = Stream(host, port)
         self.simulator_service_id = 0
-        if not Morse._asyncore_thread:
-            Morse._asyncore_thread = PollThread()
-            Morse._asyncore_thread.start()
+        if not Morse.poll_thread:
+            Morse.poll_thread = PollThread()
+            Morse.poll_thread.start()
             logger.debug("Morse thread started")
         else:
             logger.debug("Morse thread was already started")
@@ -603,14 +591,14 @@ class Morse(object):
             logger.info('Waiting for all asynchronous requests to complete...')
         self.executor.shutdown(wait = True)
         # Close all other asyncore sockets (StreanJSON)
-        if Morse._asyncore_thread:
-            Morse._asyncore_thread.syncstop(TIMEOUT)
+        if Morse.poll_thread:
+            Morse.poll_thread.syncstop(TIMEOUT)
         asyncore.close_all()
-        Morse._asyncore_thread = None # in case we want to re-create
+        Morse.poll_thread = None # in case we want to re-create
         logger.info('Done. Bye bye!')
 
     def spin(self):
-        Morse._asyncore_thread.join()
+        Morse.poll_thread.join()
 
 
     #####################################################################
